+++ dimo/api/token_exchange.py
@@ -1,4 +1,5 @@
-from dimo.constants import dimo_constants
+from constants import dimo_constants
+# from dimo.constants import dimo_constants
 
 class TokenExchange:
