+++ dimo/dimo.py
@@ -1,19 +1,38 @@
-from dimo.auth import Auth
-from dimo.device_data import DeviceData
-from dimo.device_definitions import DeviceDefinitions
-from dimo.devices import Devices
-from dimo.events import Events
-from dimo.token_exchange import TokenExchange
-from dimo.trips import Trips
-from dimo.user import User
-from dimo.valuations import Valuations
-from dimo.vehicle_signal_decoding import VehicleSignalDecoding
+# Local development imports
+from api.auth import Auth
+from api.device_data import DeviceData
+from api.device_definitions import DeviceDefinitions
+from api.devices import Devices
+from api.events import Events
+from api.token_exchange import TokenExchange
+from api.trips import Trips
+from api.user import User
+from api.valuations import Valuations
+from api.vehicle_signal_decoding import VehicleSignalDecoding
 
-from dimo.identity import Identity
-from dimo.telemetry import Telemetry
+from graphql.identity import Identity
+from graphql.telemetry import Telemetry
 
-from dimo.request import Request
-from dimo.environments import dimo_environment
+from request import Request
+from environments import dimo_environment
+
+# PyPi package imports (commented out)
+# from dimo.api.auth import Auth
+# from dimo.api.device_data import DeviceData
+# from dimo.api.device_definitions import DeviceDefinitions
+# from dimo.api.devices import Devices
+# from dimo.api.events import Events
+# from dimo.api.token_exchange import TokenExchange
+# from dimo.api.trips import Trips
+# from dimo.api.user import User
+# from dimo.api.valuations import Valuations
+# from dimo.api.vehicle_signal_decoding import VehicleSignalDecoding
+
+# from dimo.graphql.identity import Identity
+# from dimo.graphql.telemetry import Telemetry
+
+# from dimo.request import Request
+# from dimo.environments import dimo_environment
 import re
 
 class DIMO:
