+++ dimo/graphql/__init__.py
@@ -1,0 +1,4 @@
+from .identity import Identity
+from .telemetry import Telemetry
+
+__all__ = ["Identity", "Telemetry"]
